SUD-591 Build SUD-586 OTC Option Calculator неправильно подгружает и расчитывает данные SUD-585 Не работает интерполяция по соседним экспарациям в таблице имплайд волатильности 2.9 SUD-577 Не правильное отображение графиков в Volatility Manager и BO Eod Dividend Loading
git-svn-id: http://fdevstorage/svn/IVRMstandard/trunk@261 2ab15184-0dcf-aa4b-bfde-16869b366a71

+++ ETS/IVManager/ChartDlg.cpp
@@ -918,7 +918,7 @@ void CChartDlg::OnMouseUpTsChart(short Button, short Shift, long X, long Y)
 		bool bVolaInPercent = true; //m_rMgr.GetShowVolatilityInPercents();
 		long lVolaMultiplier = bVolaInPercent ? 100L : 1L;
 		
-		long dtMonth;
+		double dtMonth;
 		bool bUpdated = false;
 		if( m_nShifted )
 		{
@@ -974,7 +974,7 @@ void CChartDlg::OnDblClickTsChart()
 			bool bVolaInPercent = true; //m_rMgr.GetShowVolatilityInPercents();
 			long lVolaMultiplier = bVolaInPercent ? 100L : 1L;
 			
-			long dtMonth;
+			double dtMonth;
 			if( UpdateCustomVolaPoint( lPtIdx, ve.m_dNewVola / lVolaMultiplier, dtMonth ) )
 			{
 				SetIsDirty( TRUE );
@@ -1167,7 +1167,7 @@ BOOL CChartDlg::OnEraseBkgnd(CDC* pDC)
 
 
 #ifdef _VOLA_MANAGER
-void CChartDlg::PostNotification( UINT nMessage, CCustomVolaNotification::Type enType, long dtMonth )
+void CChartDlg::PostNotification( UINT nMessage, CCustomVolaNotification::Type enType, double dtMonth )
 {
 	CWnd *pNotificationWnd = AfxGetMainWnd();
 	if( pNotificationWnd )
